+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix missing Slack channel ID when adding emoji reaction.
+
 ## [2.0.0] - 2025-07-11
 
 ### Added